1.078

+++ src/com/teraim/vortex/dynamic/types/PhotoMeta.java
@@ -6,30 +6,30 @@ import android.util.Log;
 
 public class PhotoMeta implements Serializable {
 	private static final long serialVersionUID = -3400543797668108398L;
-	public float top=0,left=0,bottom=0,right=0;
+	public double top=0,left=0,bottom=0,right=0;
 
 
 	public PhotoMeta(String topN,String topE,String bottomN,String bottomE) {
 		try {
-			top = Float.parseFloat(topN);
-			left = Float.parseFloat(topE);
-			bottom = Float.parseFloat(bottomN);
-			right = Float.parseFloat(bottomE);
+			top = Double.parseDouble(topN);
+			left = Double.parseDouble(topE);
+			bottom = Double.parseDouble(bottomN);
+			right = Double.parseDouble(bottomE);
 		} catch (NumberFormatException e) { Log.e("vortex","non number in gis bg coordinates"); };
 
 
 	}
-	public PhotoMeta(float topN,float topE,float bottomN,float bottomE) {
+	public PhotoMeta(double topN,double topE,double bottomN,double bottomE) {
 		this.top=topN;
 		this.left=topE;
 		this.bottom=bottomN;
 		this.right=bottomE;
 	}
 	
-	public float getWidth()  {
+	public double getWidth()  {
 		return right-left;
 	}
-	public float getHeight() {
+	public double getHeight() {
 		return top-bottom;
 	}
 }
